I see this error with the app, reported by NextJS, please fix it. The er
